backend/requirements.txt
Django>=5.2.8
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
requests>=2.31.0
pymysql>=1.1.0
drf-yasg>=1.21.7
